requirements.txt
amqp==2.5.0
billiard==3.5.0.5
celery==4.1.0
certifi==2019.3.9
chardet==3.0.4
Django==1.8.2
django-haystack==2.8.1
django-redis==4.10.0
django-redis-sessions==0.5.6
django-tinymce==2.6.0
fdfs-client-py==1.2.6
get==2019.4.13
idna==2.8
itsdangerous==1.1.0
jieba==0.39
kombu==4.1.0
mutagen==1.42.0
Pillow==6.0.0
post==2019.4.13
public==2019.4.13
py3Fdfs==2.2.0
pycryptodomex==3.7.2
PyMySQL==0.9.3
python-alipay-sdk==1.10.1
pytz==2019.1
query-string==2019.4.13
redis==2.10.6
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
uWSGI==2.0.18
vine==1.3.0
Whoosh==2.7.4